aligned naming to pep 8 convention+ linting improvements
