+++ qbraid_algorithms/QTran/GateLibrary.py
@@ -439,19 +439,19 @@ class std_gates(GateLibrary):
     # ═══════════════════════════════════════════════════════════════════════════
 
     def phase(self, theta, targ):
-        """Apply phase gate: |0⟩→|0⟩, |1⟩→e^(iθ)|1⟩"""
+        """Apply phase gate: \|0⟩>\|0⟩, \|1⟩>e^(iθ)\|1⟩"""
         self.call_gate("phase", targ, phases=theta)
 
     def x(self, targ):
-        """Apply Pauli-X gate (bit flip): |0⟩→|1⟩, |1⟩→|0⟩"""
+        """Apply Pauli-X gate (bit flip): \|0⟩> \|1⟩, \|1⟩> \|0⟩"""
         self.call_gate('x', targ)
 
     def y(self, targ):
-        """Apply Pauli-Y gate: |0⟩→i|1⟩, |1⟩→-i|0⟩"""
+        """Apply Pauli-Y gate: \|0⟩>i\|1⟩, \|1⟩>-i\|0⟩"""
         self.call_gate('y', targ)
 
     def z(self, targ):
-        """Apply Pauli-Z gate (phase flip): |0⟩→|0⟩, |1⟩→-|1⟩"""
+        """Apply Pauli-Z gate (phase flip): \|0⟩> \|0⟩, \|1⟩>-\|1⟩"""
         self.call_gate('z', targ)
 
     def h(self, targ):
@@ -459,11 +459,11 @@ class std_gates(GateLibrary):
         self.call_gate('h', targ)
 
     def s(self, targ):
-        """Apply S gate (phase): |1⟩→i|1⟩"""
+        """Apply S gate (phase): \|1⟩>i\|1⟩"""
         self.call_gate('s', targ)
 
     def sdg(self, targ):
-        """Apply S-dagger gate (inverse phase): |1⟩→-i|1⟩"""
+        """Apply S-dagger gate (inverse phase): \|1⟩>-i\|1⟩"""
         self.call_gate('sdg', targ)
 
     def sx(self, targ):
