+++ qbraid_algorithms/QTran/__init__.py
@@ -30,7 +30,7 @@ Functions
     
 """
 from .GateLibrary import GateLibrary, std_gates
-from .QasmBuilder import FileBuilder, GateBuilder, IncludeBuilder, QasmBuilder
 from .ModuleLoader import qasm_pipe
+from .QasmBuilder import FileBuilder, GateBuilder, IncludeBuilder, QasmBuilder
 
 __all__ = ['FileBuilder', 'QasmBuilder','GateBuilder','IncludeBuilder','GateLibrary','std_gates','qasm_pipe']
