+++ qbraid_algorithms/qft/QFTLibrary.py
@@ -15,7 +15,7 @@
 # from QasmBuilder import FileBuilder, QasmBuilder, GateBuilder
 import string
 
-from qbraid_algorithms.QTran import GateLibrary, GateBuilder, std_gates
+from qbraid_algorithms.QTran import GateBuilder, GateLibrary, std_gates
 
 
 class QFTLibrary(GateLibrary):
